refactor create_reports to reports
